+++ scripts/install.ba.sh
@@ -12,12 +12,14 @@ bash ${scripts_path}/pip/upgrade.ba.sh
 # pipx > uv
 bash ${scripts_path}/pipx/install.ba.sh
 
-# dependencies & tools
+# uv > tools
 bash ${scripts_path}/uv/tool/install.ba.sh
-bash ${scripts_path}/uv/sync.ba.sh
+
+# dependencies
+bash ${scripts_path}/pip/install.ba.sh
 
 # jupyter .venv
-bash ${scripts_path}/ipykernel/install.ba.sh
+# bash ${scripts_path}/ipykernel/install.ba.sh # TODO?
 
 set +x
 set +e