+++ scripts/uv/tool/install.ba.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+
+set -e
+
+filename="requirements.uv_tool.txt"
+if [[ -z $(grep '[^[:space:]]' $filename) ]]; then
+  echo "${filename} is empty, skipping..."
+  exit 0
+fi
+
+set -v
+
+cat ${filename} | sed 's/.*/"&"/' | xargs -n 1 uv tool install
+
+set +v
+set +e