+++ HISTORY.md
@@ -6,6 +6,26 @@
 
 1. `$> pipx install uv`
 1. `$> uv self update`
+
+#### python
+
+1. `$> uv python install 3.12.7`
+1. `$> uv python pin 3.12.7`
+
+#### venv
+
+1. `$> uv venv --python 3.12.7`
+
+#### pyproject.toml
+
 1. `$> uv init`
 1. `$> uv add {dependency-name}`
+
+#### tools
+
 1. `$> uv tool add {tool-name}`
+
+#### References
+
+- [uv](https://github.com/astral-sh/uv)
+- [Installing uv](https://docs.astral.sh/uv/getting-started/installation/#pypi)